Updates Search command to slash command
